requirements.txt
gensim
tqdm
pytorch-crf